# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "CurificBackend Copy",
		"contact": {},
		"version": "1.0"
	},
	"servers": [
		{
			"url": "http://65.2.142.187:5001/api/v1",
			"variables": {}
		}
	],
	"paths": {
...
		"/timeSlot/assign": {
			"put": {
				"tags": ["Dr &customer"],
				"summary": "AssignTimeSlot",
				"description": "Assign Time Slot",
				"operationId": "AssignTimeSlot",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AssignTimeSlotRequest"
							},
							"example": {
								"drId": "62d95849c1da0d310474a30d",
								"timeSlot": [
									"62d7a59a9492b434857b8e4a",
									"62d7a5b99492b434857b8e50",
									"62d7a5cf9492b434857b8e56",
									"62d7a5de9492b434857b8e5c",
									"62d7a64a9492b434857b8e62",
									"62d7a65d9492b434857b8e68",
									"62d7a66c9492b434857b8e6e",
									"62d7a67a9492b434857b8e74"
								]
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AssignTimeSlot"
								},
								"example": {
									"success": true,
									"msg": "Time Slot assign successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false
			}
		},
...
		"/notificationType/create": {
			"post": {
				"tags": ["NotificationType"],
				"summary": "createnotificationtype",
				"description": "create notification type",
				"operationId": "createnotificationtype",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createnotificationtyperequest"
							},
							"example": {
								"type": "SMS",
								"message": "Hello, This is gental reminder for you appointment"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
		"/notificationType/update/{id}": {
			"put": {
				"tags": ["NotificationType"],
				"summary": "updatenotificationtype",
				"description": "update notification type",
				"operationId": "updatenotificationtype",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updatenotificationtyperequest"
							},
							"example": {
								"type": "SMS",
								"message": "Hello, This is gental reminder for you appointment !"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
		"/notificationType/delete/{id}": {
			"delete": {
				"tags": ["NotificationType"],
				"summary": "deletenotificationtype",
				"description": "delete notification type",
				"operationId": "deletenotificationtype",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
...
		"/timeSlot/create": {
			"post": {
				"tags": ["Time Slot"],
				"summary": "CreatetimeSlot",
				"description": "Create time Slot",
				"operationId": "CreatetimeSlot",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatetimeSlotRequest"
							},
							"example": {
								"startTime": "10:00 A",
								"endTime": "10:30 A"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m20011"
								},
								"example": {
									"success": true,
									"msg": "Time Slot created successfully",
									"data": {
										"deletedAt": null,
										"deletedBy": null,
										"_id": "62ce6bee37e91b0a80fddb2b",
										"startTime": "10:00 A",
										"endTime": "10:30 A",
										"createdAt": "2022-07-13T06:53:34.046Z",
										"updatedAt": "2022-07-13T06:53:34.046Z",
										"__v": 0
									}
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/timeSlot/update/{id}": {
			"put": {
				"tags": ["Time Slot"],
				"summary": "UpdatetimeSlot",
				"description": "Update time Slot",
				"operationId": "UpdatetimeSlot",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatetimeSlotRequest"
							},
							"example": {
								"startTime": "10:10 A",
								"endTime": "10:30 A"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2002"
								},
								"example": {
									"success": true,
									"msg": "Time Slot update successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/timeSlot/delete/{id}": {
			"delete": {
				"tags": ["Time Slot"],
				"summary": "DeletetimeSlot",
				"description": "Delete time Slot",
				"operationId": "DeletetimeSlot",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2002"
								},
								"example": {
									"success": true,
									"msg": "Time Slot delete successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
...
		"/package/create": {
			"post": {
				"tags": ["Package Management"],
				"summary": "CreatePackage",
				"description": "Create Package",
				"operationId": "CreatePackage",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePackageRequest"
							},
							"example": {
								"name": "test package11",
								"description": "description package",
								"priceByMonth": 10,
								"priceByYear": 100,
								"shortDescription": "test"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2002"
								},
								"example": {
									"success": true,
									"msg": "Package created successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/package/update/{id}": {
			"put": {
				"tags": ["Package Management"],
				"summary": "UpdatePackage",
				"description": "Update Package",
				"operationId": "UpdatePackage",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePackageRequest"
							},
							"example": {
								"name": "test package",
								"description": "description package",
								"priceByMonth": 10,
								"priceByYear": 100,
								"shortDescription": "aaa",
								"image": "/images/package/1658486962368.png"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2002"
								},
								"example": {
									"success": true,
									"msg": "Package update successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/package/delete/{id}": {
			"delete": {
				"tags": ["Package Management"],
				"summary": "deletepackage",
				"description": "delete package",
				"operationId": "deletepackage",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2002"
								},
								"example": {
									"success": true,
									"msg": "Package delete successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"AssignTimeSlotRequest": {
				"title": "AssignTimeSlotRequest",
				"required": ["drId", "timeSlot"],
				"type": "object",
				"properties": {
					"drId": {
						"type": "string"
					},
					"timeSlot": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"drId": "62d95849c1da0d310474a30d",
					"timeSlot": [
						"62d7a59a9492b434857b8e4a",
						"62d7a5b99492b434857b8e50",
						"62d7a5cf9492b434857b8e56",
						"62d7a5de9492b434857b8e5c",
						"62d7a64a9492b434857b8e62",
						"62d7a65d9492b434857b8e68",
						"62d7a66c9492b434857b8e6e",
						"62d7a67a9492b434857b8e74"
					]
				}
			},
			"AssignTimeSlot": {
				"title": "AssignTimeSlot",
				"required": ["success", "msg", "data"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"data": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"success": true,
					"msg": "Time Slot assign successfully",
					"data": []
				}
			},
...
			"m2002": {
				"title": "m2002",
				"required": ["success", "msg", "data"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"data": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"success": true,
					"msg": "Doctor delete successfully",
					"data": []
				}
			},
...
			"createnotificationtyperequest": {
				"title": "createnotificationtyperequest",
				"required": ["type", "message"],
				"type": "object",
				"properties": {
					"type": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"example": {
					"type": "SMS",
					"message": "Hello, This is gental reminder for you appointment"
				}
			},
			"updatenotificationtyperequest": {
				"title": "updatenotificationtyperequest",
				"required": ["type", "message"],
				"type": "object",
				"properties": {
					"type": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"example": {
					"type": "SMS",
					"message": "Hello, This is gental reminder for you appointment !"
				}
			},
...
			"CreatetimeSlotRequest": {
				"title": "CreatetimeSlotRequest",
				"required": ["startTime", "endTime"],
				"type": "object",
				"properties": {
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					}
				},
				"example": {
					"startTime": "10:00 A",
					"endTime": "10:30 A"
				}
			},
			"m20011": {
				"title": "m20011",
				"required": ["success", "msg", "data"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/Data4"
					}
				},
				"example": {
					"success": true,
					"msg": "Time Slot created successfully",
					"data": {
						"deletedAt": null,
						"deletedBy": null,
						"_id": "62ce6bee37e91b0a80fddb2b",
						"startTime": "10:00 A",
						"endTime": "10:30 A",
						"createdAt": "2022-07-13T06:53:34.046Z",
						"updatedAt": "2022-07-13T06:53:34.046Z",
						"__v": 0
					}
				}
			},
			"Data4": {
				"title": "Data4",
				"required": [
					"deletedAt",
					"deletedBy",
					"_id",
					"startTime",
					"endTime",
					"createdAt",
					"updatedAt",
					"__v"
				],
				"type": "object",
				"properties": {
					"deletedAt": {
						"type": "string",
						"nullable": true
					},
					"deletedBy": {
						"type": "string",
						"nullable": true
					},
					"_id": {
						"type": "string"
					},
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"__v": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"deletedAt": null,
					"deletedBy": null,
					"_id": "62ce6bee37e91b0a80fddb2b",
					"startTime": "10:00 A",
					"endTime": "10:30 A",
					"createdAt": "2022-07-13T06:53:34.046Z",
					"updatedAt": "2022-07-13T06:53:34.046Z",
					"__v": 0
				}
			},
			"UpdatetimeSlotRequest": {
				"title": "UpdatetimeSlotRequest",
				"required": ["startTime", "endTime"],
				"type": "object",
				"properties": {
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					}
				},
				"example": {
					"startTime": "10:10 A",
					"endTime": "10:30 A"
				}
			},
...
			"CreatePackageRequest": {
				"title": "CreatePackageRequest",
				"required": [
					"name",
					"description",
					"priceByMonth",
					"priceByYear",
					"shortDescription"
				],
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"priceByMonth": {
						"type": "integer",
						"format": "int32"
					},
					"priceByYear": {
						"type": "integer",
						"format": "int32"
					},
					"shortDescription": {
						"type": "string"
					}
				},
				"example": {
					"name": "test package11",
					"description": "description package",
					"priceByMonth": 10,
					"priceByYear": 100,
					"shortDescription": "test"
				}
			},
			"UpdatePackageRequest": {
				"title": "UpdatePackageRequest",
				"required": [
					"name",
					"description",
					"priceByMonth",
					"priceByYear",
					"shortDescription",
					"image"
				],
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"priceByMonth": {
						"type": "integer",
						"format": "int32"
					},
					"priceByYear": {
						"type": "integer",
						"format": "int32"
					},
					"shortDescription": {
						"type": "string"
					},
					"image": {
						"type": "string"
					}
				},
				"example": {
					"name": "test package",
					"description": "description package",
					"priceByMonth": 10,
					"priceByYear": 100,
					"shortDescription": "aaa",
					"image": "/images/package/1658486962368.png"
				}
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"security": [],
...
}